requirements.txt
gdown
tensorflow
matplotlib
numpy
Pillow
twdm